+++ packages/next/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @scribe-atp/next
 
+## 1.1.0
+
+### Minor Changes
+
+- Export `buildCanonicalUrl` from `@scribe-atp/core`. Add `articleMetadata` and `siteMetadata` to `@scribe-atp/next`, returning Next.js `Metadata` objects for use in `generateMetadata`. Add `articleSeoMeta` and `siteSeoMeta` to `@scribe-atp/nuxt`, returning objects shaped for Nuxt's `useSeoMeta` composable.
+
+### Patch Changes
+
+- Updated dependencies
+  - @scribe-atp/core@3.3.1
+
 ## 1.0.1
 
 ### Patch Changes